mejora de autorization de login habia un error de inicio solucionado

+++ client/src/services/api.ts
@@ -5,7 +5,7 @@ import axios from "axios";
 
 
 const api = axios.create({
-    baseURL: import.meta.env.VITE_API_URL   // ← variable de entorno super importante mejorar
+    baseURL: import.meta.env.VITE_API_URL   
 
 })
 
@@ -25,17 +25,19 @@ api.interceptors.request.use(config => {
 
 
 api.interceptors.response.use(
+
     (response) => response,
     (error) => {
-        if (error.response?.status === 401) {
+
+        const isLoginRequest = error.config?.url?.includes('/login')
+
+        if (error.response?.status === 401 && !isLoginRequest) {
             localStorage.clear()
             router.push('/login')
         }
         return Promise.reject(error)
     }
 )
-
-
 
 
 export default api
